+++ src/Math/AxisAngle.h
@@ -15,12 +15,12 @@ namespace Raycer
 	{
 	public:
 
-		explicit AxisAngle(const Vector3& axis = Vector3::UP, double angle = 0.0);
+		explicit AxisAngle(const Vector3& axis = Vector3::UP, float angle = 0.0f);
 
 		Matrix4x4 toMatrix4x4() const;
 
 		Vector3 axis;
-		double angle; // degrees
+		float angle; // degrees
 
 	private:
